+++ lamb/management/commands/alchemy_create.py
@@ -7,8 +7,7 @@ from sqlalchemy.exc import DBAPIError, SQLAlchemyError
 from sqlalchemy.ext.compiler import compiles
 from sqlalchemy.schema import DropSequence, DropTable
 
-from lamb.db.session import metadata
-from lamb.management.base import LambLoglevelMixin
+from lamb.management.base import LambCommandMixin
 from lamb.utils.core import compact
 
 logger = logging.getLogger(__name__)
@@ -29,7 +28,7 @@ def _compile_drop_type(element, compiler, **_):
     return compiler.visit_drop_enum_type(element) + " CASCADE"
 
 
-class Command(LambLoglevelMixin, LabelCommand):
+class Command(LambCommandMixin, LabelCommand):
     help = "Creates database table for provided modules"  # noqa: A003
 
     def add_arguments(self, parser):
@@ -51,7 +50,9 @@ class Command(LambLoglevelMixin, LabelCommand):
 
     def handle_label(self, label, **options):
         try:
-            import_module(label)
+            metadata = self.db_metadata
+            import_module(label)  # loading models
+
             if options["exclude_tables"] is not None:
                 exclude_tables = options["exclude_tables"].split(",")
                 tables = [v for k, v in metadata.tables.items() if k not in exclude_tables]